==> 1.Data_cleaning.ipynb <==
## 1. Data Cleaning and processing

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib as mtplt
import datetime as dat
from datetime import datetime

#Importing my files
df_import_1 = pd.read_csv('Energy2017.csv')
df_import_2 = pd.read_csv('Energy2018.csv')
df_import_3 = pd.read_csv('Meteo_data_2017_2018_2019.dat')

#Making "RAW" copies for working
df_1 = df_import_1.copy(deep = True)
df_2 = df_import_2.copy(deep = True) 
df_3 = df_import_3.copy(deep = True) 

df_3 = df_3 [~df_3['yyyy-mm-dd hh:mm:ss'].str.contains('2019', na=False)]
df_3.info()

#Now cutting the redundant lines and columns
df_1 = df_1.iloc[:8760]
df_2 = df_2.iloc[:8760]
df_1 = df_1.drop (columns = ['Duration','Date_end']) 
df_2 = df_2.drop (columns = ['Duration','Date_end']) 
df_1.rename(columns = {'Power_kW': 'Power_kWh'}, inplace = True)
df_2.rename(columns = {'Power_kW': 'Power_kWh'}, inplace = True)

df_1['Date'] = sorted(pd.to_datetime(df_1['Date_start']))
df_2['Date'] = sorted(pd.to_datetime(df_2['Date_start']))
#adding 'sorted' before the parenthesis is important to avoid switching between the days and months


df_3['pres_mbar'] = df_3['pres_mbar'].astype(float)
df_3 = df_3.drop (columns = ['windSpeed_m/s','windGust_m/s','rain_mm/h','rain_day'])

# Merging the files into a one file
df = df_1.merge(df_2,how = 'outer')
df.tail()
df.columns

#Creating the weekday column
df['day_of_week'] = df['Date'].dt.weekday
df['day_of_week_name'] = df['Date'].dt.day_name

# Changing the numbering so that it starts with Monday = 1, to Sunday = 7

df['Weekday'] = df['day_of_week'] + 1 
df = df.drop (columns = ['day_of_week'])
df[0:10]

#Adding the holidays column

holiday = ['2017-01-01', '2017-04-14', '2017-04-16', '2017-04-25', '2017-05-01', '2017-06-10', '2017-06-15', 
                  '2017-08-15', '2017-10-05', '2017-11-01', '2017-12-01', '2017-12-08', '2017-12-25','2018-01-01', '2018-03-30', '2018-04-01', '2018-04-25', '2018-05-01', '2018-05-31', '2018-06-10',  
                  '2018-08-15', '2018-10-05', '2018-11-01', '2018-12-01', '2018-12-08', '2018-12-25',('2017-02-16', '2017-02-20'), ('2017-02-27', '2017-03-01'), ('2017-04-10', '2017-04-15'),
                   ('2017-08-01', '2017-09-11'), ('2017-12-23', '2017-12-30'),('2018-02-15', '2018-02-19'), ('2018-03-26', '2018-03-31'), ('2018-08-01', '2018-09-10'), 
                  ('2018-12-22', '2019-01-01')]

df['Holiday'] = [df['Date'][i].strftime("%Y-%m-%d") in holiday for i in range(df.shape[0])]

df.Holiday = df.Holiday.replace({True: 1, False: 0}) # changing the true and false booleans into 0 and 1 values
df['Holiday'] = df['Holiday'].astype(float) #changing 1 and 0 values to floats to avoid errors in the feature selection part

#now if the given day is a holiday, we will see 1.0 in the holiday column

df_3 = df_3.rename(columns = {'yyyy-mm-dd hh:mm:ss':'Date'})
df_3.rename(columns = {'solarRad_W/m2': 'Solar'}, inplace = True)

df_3['Date'] = pd.to_datetime(df_3['Date']) # changing into the datetime format

df_3 = df_3.resample('60min', on = 'Date').mean() # down sampling the weather file
df_3.head()

df = df.merge(df_3, how='left', on=['Date']) # merging the power and weather files
df.info()

df = df.drop (columns = ['Current1_Ah','Current2_Ah','Current3_Ah','HR','pres_mbar','Date_start'])

df = df.interpolate(method='linear', limit_direction ='forward') # filling in the empty records in the weather file columns (around 2000 data missing - filling based on the interpolation)

df[df.isnull().any(axis = 'columns')] # checking if i have any NaN entries

df['hour'] = df['Date'].dt.hour
df.dtypes #checking if everything is correct

#changing ints into floats
df['Weekday'] = df['Weekday'].astype(float)
df['hour'] = df['hour'].astype(float)
df = df.drop (columns = ['Total_Ah'])

df.info()

df['year'] = df['Date'].dt.year
df['month'] = df['Date'].dt.month

# Save clean dataset
df.to_csv('clean_df.csv')
