==> 2.Explanatory_data_analysis.ipynb <==
## 2. Explanatory Data Analysis

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib as mtplt
import datetime as dat
from datetime import datetime
import matplotlib.ticker as ticker

df = pd.read_csv('clean_df.csv')
df['Date'] = sorted(pd.to_datetime(df['Date']))
df.describe() # rows count, mean, std, min, percentiles, max. Getting the sense if the data have outliers, extreme values, variation, etc.

#Checking for outliers
df.sort_values(by = 'Power_kWh', ascending = False)[:5]

df.sort_values(by = 'Power_kWh', ascending = True)[:5]

df.sort_values(by = 'temp_C', ascending = False)[:5]

df.sort_values(by = 'temp_C', ascending = True)[:5]

plt.hist(df ['Power_kWh'], 20) ## Note: identifying outliers and a possibility for peak shaving?

data, bins, patches = plt.hist (df['temp_C'], 20, ec = 'k') 
cmap = plt.get_cmap('jet')
low = cmap (0.6) 
medium = cmap (0.8)
high = cmap (0.9)

for i in range(0,6): 
    patches[i].set_facecolor(low)
for i in range(6,13): 
    patches[i].set_facecolor(medium)
for i in range(13,20):
    patches[i].set_facecolor(high)

handles = [mtplt. patches.Rectangle ((0,0), 1, 1, color = c, ec = "k") for c in [low, medium, high]]
labels = ["low","medium", "high"]
plt.legend (handles, labels)
    
plt.xlabel ("Temperature, C", fontsize = 12)
plt.ylabel ("Number of occurances", fontsize = 12)
plt.xticks (fontsize = 12)
plt.yticks (fontsize = 12)

plt.title ('Temperature Histogram')

plt.show

## Plot for the selected time series

df_plot_time = df.copy (deep = True)
df_plot_time = df_plot_time.set_index ('Date', drop = False)
df_plot_time.head()

fig, ax = plt.subplots()
fig.set_size_inches(20,10)

ax.xaxis.set_major_locator (ticker.MultipleLocator(60))
ax.xaxis.set_tick_params (which = 'major', pad = 5, labelrotation = 50)

x_2017 = df_plot_time ['Date'] ['2017-01-01 00:00:00' : '2017-01-20 23:00:00']
x_2017 = x_2017.dt.strftime ('%Y-%b-%d-%H h')

y_2017 = df_plot_time ['Power_kWh'][plot_start : plot_end] # select data slice

plt.plot (x_values, y_values2017, '-o', color = 'blue', # x axis laels; data; symbol type *try '-p'; line color;
         markersize = 10, linewidth = 1, # point size; line thickness;
         markerfacecolor = 'red', # color inside the point
         markeredgecolor = 'black', # color of edge
         markeredgewidth = 3)

## For two series

fig, ax = plt.subplots()
fig.set_size_inches(20,10)

ax.xaxis.set_major_locator (ticker.MultipleLocator(60))
ax.xaxis.set_tick_params(which = 'major', pad = 5, labelrotation = 50)

x_2017 = df_plot_time ['Date'] ['2017-01-01 00:00:00' : '2017-01-15 23:00:00']
x_2017 = x_2017.dt.strftime ('%b-%d-%H h')
y_2017 = df_plot_time ['Power_kWh']['2017-01-01 00:00:00' : '2017-01-15 23:00:00']

x_2018 = df_plot_time ['Date'] ['2018-01-01 00:00:00' : '2018-01-15 23:00:00']
x_2018 = x_2018.dt.strftime ('%b-%d-%H h')
y_2018 = df_plot_time ['Power_kWh']['2018-01-01 00:00:00' : '2018-01-15 23:00:00']

plt.plot (x_2017, y_2017, '-p', color = 'black',
         markersize = 2, linewidth = 0.5,
         markerfacecolor = 'white',
         markeredgecolor = 'blue',
         markeredgewidth = 2)

plt.plot(x_2018, y_2018, '-p', color = 'red',
         markersize=2, linewidth=0.5,
         markerfacecolor='white',
         markeredgecolor='red',
         markeredgewidth=2)
