==> 3.Clustering.ipynb <==
## Clustering

import pandas as pd
from pandas import DataFrame
import numpy as np
import matplotlib.pyplot as plt
import matplotlib as mtplt
import datetime 

df = pd.read_csv('clean_df.csv')
df['Date'] = sorted(pd.to_datetime(df['Date']))

#Clustering
from sklearn.cluster import KMeans

## Power and hour

kmeans = KMeans(n_clusters=3)
cluster = DataFrame(df,columns=['hour','Power_kWh'])

kmeans.fit(cluster)
print(kmeans.cluster_centers_)

y_km = kmeans.fit_predict(cluster)
print(y_km)

plt.scatter(cluster.hour[y_km ==0], cluster.Power_kWh[y_km == 0], s=10, c='red')
plt.scatter(cluster.hour[y_km ==1], cluster.Power_kWh[y_km == 1], s=10, c='black')
plt.scatter(cluster.hour[y_km ==2], cluster.Power_kWh[y_km == 2], s=10, c='green')

## Power and weekday

kmeans = KMeans(n_clusters=3)
cluster = DataFrame(df,columns=['Weekday','Power_kWh'])

kmeans.fit(cluster)
print(kmeans.cluster_centers_)

y_km = kmeans.fit_predict(cluster)
print(y_km)

plt.scatter(cluster.Weekday[y_km ==0], cluster.Power_kWh[y_km == 0], s=10, c='red')
plt.scatter(cluster.Weekday[y_km ==1], cluster.Power_kWh[y_km == 1], s=10, c='black')
plt.scatter(cluster.Weekday[y_km ==2], cluster.Power_kWh[y_km == 2], s=10, c='green')

## Power and temperature

kmeans = KMeans(n_clusters=3)
cluster = DataFrame(df,columns=['temp_C','Power_kWh'])

kmeans.fit(cluster)
print(kmeans.cluster_centers_)

y_km = kmeans.fit_predict(cluster)
print(y_km)

plt.scatter(cluster.temp_C[y_km ==0], cluster.Power_kWh[y_km == 0], s=10, c='red')
plt.scatter(cluster.temp_C[y_km ==1], cluster.Power_kWh[y_km == 1], s=10, c='black')
plt.scatter(cluster.temp_C[y_km ==2], cluster.Power_kWh[y_km == 2], s=10, c='green')

## Power and Solar irradiaion 


kmeans = KMeans(n_clusters=3)
cluster = DataFrame(df,columns=['Solar','Power_kWh'])

kmeans.fit(cluster)
print(kmeans.cluster_centers_)

y_km = kmeans.fit_predict(cluster)
print(y_km)

plt.scatter(cluster.Solar[y_km ==0], cluster.Power_kWh[y_km == 0], s=10, c='red')
plt.scatter(cluster.Solar[y_km ==1], cluster.Power_kWh[y_km == 1], s=10, c='black')
plt.scatter(cluster.Solar[y_km ==2], cluster.Power_kWh[y_km == 2], s=10, c='green')
