==> 4.Feature_selection.ipynb <==
## 4. Feature Selection

import pandas as pd
from pandas import DataFrame
import numpy as np
import matplotlib.pyplot as plt
import matplotlib as mtplt
import datetime 

df = pd.read_csv('clean_df.csv')
df['Date'] = sorted(pd.to_datetime(df['Date']))

df_feature_selection = df.drop (columns = ['day_of_week_name','Unnamed: 0','month','year']) # removing the column with str, cause it causes problems with feature selection algorithms
df_feature_selection = df_feature_selection.set_index('Date') #make 'datetime' into index
df_feature_selection['Power-1']=df_feature_selection['Power_kWh'].shift(1) # Previous hour consumption
df_feature_selection['Power-2']=df_feature_selection['Power_kWh'].shift(2) # Previous hour consumption
df_feature_selection

df_feature_selection = df_feature_selection.dropna()
df_feature_selection.head()

df_feature_selection.to_csv('df_feature_selection.csv')

# Define input and outputs
X=df_feature_selection.values
Y=X[:,0] # prediction variable
X=X[:,[1,2,3,4,5,6,7]] # features that contribute to my predicted variable

## K-Best method

from sklearn.feature_selection import SelectKBest
from sklearn.feature_selection import f_regression

features=SelectKBest(k=3,score_func=f_regression)
fit=features.fit(X,Y)
print(fit.scores_)
features_results=fit.transform(X)
print(features_results) #weekday, holiday, temp_C, Solar, hour, Power-1, Power-2 ---- order of the features

# Importance: Power-1, Power-2

## Random forest regressor

from sklearn.ensemble import RandomForestRegressor
model = RandomForestRegressor()
model.fit(X, Y)
print(model.feature_importances_) #weekday, holiday, temp_C, Solar, hour, Power-1, Power-2 ---- order of the featuers

# Importance: Power -1, hour

## Recursive Feature Elimination

#Recursive Feature Elimination

from sklearn.feature_selection import RFE
from sklearn.linear_model import LinearRegression

model=LinearRegression() # LinearRegression Model as Estimator
rfe1=RFE(model,2)# using 2 features
rfe2=RFE(model,3) # using 3 features

fit1=rfe1.fit(X,Y)
fit2=rfe2.fit(X,Y)

print( "Feature Ranking (Linear Model, 2 features): %s" % (fit1.ranking_)) 
print( "Feature Ranking (Linear Model, 3 features): %s" % (fit2.ranking_)) 

#Most important data: Power -1, hour, Power-2
