==> 5.Forecasting_models.ipynb <==
## 5. Forecast

import pandas as pd
from pandas import DataFrame
import numpy as np
import matplotlib.pyplot as plt
import matplotlib as mtplt
import datetime 

df = pd.read_csv('df_feature_selection.csv')
df['Date'] = sorted(pd.to_datetime(df['Date']))
df = df.set_index('Date')
df

from sklearn.model_selection import train_test_split
from sklearn import  metrics
from sklearn import  linear_model
from sklearn.ensemble import RandomForestRegressor

X=df.values
Y=X[:,0]
X=X[:,[1,5,6,7]] # Data from the feature selection (Power-1, Power-2, hour)

X_train, X_test, y_train, y_test = train_test_split(X,Y)
print(X_train)
print(y_train)

## Linear Regression

from sklearn import  linear_model

regr = linear_model.LinearRegression()

regr.fit(X_train,y_train)

y_pred_LR = regr.predict(X_test)

plt.plot(y_test[1:200],  c='black')
plt.plot(y_pred_LR[1:200], c = 'red') #red is the forecast
plt.show()
plt.scatter(y_test,y_pred_LR, s=10, c='green')

#Evaluate errors
MAE_LR=metrics.mean_absolute_error(y_test,y_pred_LR) 
MSE_LR=metrics.mean_squared_error(y_test,y_pred_LR)  
RMSE_LR= np.sqrt(metrics.mean_squared_error(y_test,y_pred_LR))
cvRMSE_LR=RMSE_LR/np.mean(y_test) #error in %
print(MAE_LR, MSE_LR, RMSE_LR,cvRMSE_LR)

## Support Vector Machine

from sklearn.preprocessing import StandardScaler
from sklearn.svm import SVR

sc_X = StandardScaler()
sc_y = StandardScaler()
X_train = sc_X.fit_transform(X_train)
y_train = sc_y.fit_transform(y_train.reshape(-1,1))


regr = SVR(kernel='rbf')
regr.fit(X_train,y_train)

y_pred_SVR = regr.predict(sc_X.fit_transform(X_test))
y_test_SVR = sc_y.fit_transform(y_test.reshape(-1,1))
y_pred_SVR2 = sc_y.inverse_transform(y_pred_SVR)

plt.plot(y_test_SVR[1:200], c = 'black')
plt.plot(y_pred_SVR[1:200], c = 'red') # red is the forecast
plt.show()
plt.plot(y_test[1:200], c='black')
plt.plot(y_pred_SVR2[1:200], c='red')

plt.scatter(y_test_SVR,y_pred_SVR2, s=10, c='green')
plt.show()

MAE_SVR=metrics.mean_absolute_error(y_test,y_pred_SVR2) 
MSE_SVR=metrics.mean_squared_error(y_test,y_pred_SVR2)  
RMSE_SVR= np.sqrt(metrics.mean_squared_error(y_test,y_pred_SVR2))
cvRMSE_SVR=RMSE_SVR/np.mean(y_test)
print(MAE_SVR, MSE_SVR, RMSE_SVR,cvRMSE_SVR)

## Choosing the model

Support Vector Machine

df_predicted = pd.DataFrame({'Predicted':y_pred_SVR2})
df_predicted['Date'] = pd.date_range('1/1/2019', periods = len(y_pred_SVR2), freq ='H') 
df_predicted

#Save the forecast
df_predicted.to_csv('df_forecast.csv')
